% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frames.R
\name{frame_1}
\alias{frame_1}
\title{Put Map In A Custom Frame}
\usage{
frame_1(
  map,
  title_text = "Title",
  subtitle_text = "Subtitle",
  title_font = "Brush Script MT",
  subtitle_font = "Trebuchet MS"
)
}
\description{
NEED TO ADD DESCRIPTION
}
