% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frames.R
\name{map_frame}
\alias{map_frame}
\title{Put Map In A Custom Frame}
\usage{
map_frame(
  map,
  title_text = "Title",
  subtitle_text = "",
  title_font = "Brush Script MT",
  subtitle_font = "Trebuchet MS",
  frame_width = "100\%",
  frame_height = 780,
  mask = "circular_mask"
)
}
\arguments{
\item{map}{a map you created with mapBliss or leaflet}

\item{title_text}{Title text}

\item{subtitle_text}{Subtitle text}

\item{title_font}{Title font}

\item{subtitle_font}{Subtitle font}

\item{frame_width}{Set by default to 100 percent. It is recommended not to change this.}

\item{frame_height}{Set by default to 780 so as to fill the default shiny screen. Update according to your needs.}

\item{mask}{The function will look in masks/ for a PNG file with the same name. Defaults to circular_mask.}
}
\description{
Put your maps in a customized frame with `map_frame()` functions.
}
\examples{

plot_city_view("Jersalem, IL") |>
  map_frame(title_text="Jerusalem", subtitle_text="City of Gold", subtitle_font="Brush Script MT")

plot_city_view("Amsterdam") |>
  map_frame(title_text="Amsterdam", mask="house_mask")
}
