% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_city_view.R
\name{plot_city_view}
\alias{plot_city_view}
\title{Plot A City View Map}
\usage{
plot_city_view(
  city,
  mapBoxTemplate = "//{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
  zoomControl = 0.1
)
}
\arguments{
\item{city}{the city}
}
\description{
Documentation is needed
}
\examples{

plot_city_view("New York City")

plot_city_view("Paris")
}
