% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flights.R
\name{plot_flights}
\alias{plot_flights}
\title{Plot Flight Paths}
\usage{
plot_flights(
  addresses,
  colour = "black",
  opacity = 1,
  weight = 1,
  radius = 2,
  label_text = addresses,
  label_position = "bottom",
  font = "Lucida Console",
  font_weight = "bold",
  font_size = "14px",
  text_indent = "15px",
  mapBoxTemplate = "//{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
  nCurves = 100
)
}
\description{
Plot multiple flights with `plot_flights`. This function produces a html object so in order to save the image as an .svg (for printable visuals) check out the `save_map_svg()`.
}
\examples{
plot_flights()
}
