% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hybrid_route.R
\name{plot_hybrid_route}
\alias{plot_hybrid_route}
\title{Plot Hybrid Route}
\usage{
plot_hybrid_route(
  addresses,
  how = c("car", ...),
  colour = "black",
  opacity = 1,
  weight = 1,
  radius = 2,
  label_text = addresses,
  label_position = "bottom",
  font = "Lucida Console",
  font_weight = "bold",
  font_size = "14px",
  text_indent = "15px",
  mapBoxTemplate = "//{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
)
}
\description{
Documentation is needed
}
