% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_route.R
\name{plot_route}
\alias{plot_route}
\title{Plot route}
\usage{
plot_route(
  to,
  from,
  how = "car",
  colour = "black",
  opacity = 1,
  weight = 1,
  radius = 2,
  label_text = c(to, from),
  label_position = "bottom",
  provider = providers$CartoDB.PositronNoLabels,
  font = "Lucida Console",
  font_weight = "bold",
  font_size = "14px",
  text_indent = "15px",
  saturation = 0,
  mapBoxTemplate = "//{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
)
}
\description{
Documentation is needed
}
\details{
For more information on how to integrate the MapBox API to make beautiful maps, check out [this blog](https://bensstats.wordpress.com/2021/10/25/robservations-16-using-the-mapbox-api-with-leaflet/).
}
