% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_route_flex.R
\name{plot_route_flex}
\alias{plot_route_flex}
\title{Plot Road Trip Route With More Flexibility}
\usage{
plot_route_flex(
  addresses,
  how = c("car", "bike", "foot"),
  colour = "black",
  opacity = 1,
  weight = 1,
  radius = 2,
  label_text = addresses,
  label_position = "bottom",
  font = "Lucida Console",
  font_weight = "bold",
  font_size = "14px",
  text_indent = "15px",
  mapBoxTemplate = "//{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
  zoomControl = c(0.1, 0.1, -0.1, -0.1)
)
}
\arguments{
\item{addresses}{the stops on your route}

\item{how}{routing profile(s) to use, e.g. "car", "bike" or "foot" (when using the routing.openstreetmap.de test server). n-1 routing profiles are required for for n addresses}

\item{colour}{what colour you want the route line to be colored}

\item{opacity}{line opacity - a value between 0 and 1}

\item{weight}{Line thickness}

\item{radius}{Point size}

\item{label_text}{Optional. Alternative text to display as the address labels.}

\item{label_position}{where to place teach of the labels relative to the points ("bottom", "top", "left", "right")}

\item{font}{font-family css property}

\item{font_weight}{font-weight css property}

\item{font_size}{font-size css property}

\item{text_indent}{text indent css property}

\item{mapBoxTemplate}{the MapBox template you want to use.}

\item{zoomControl}{The zoom control for your bounding box. Default format is `c(lng1,lat1,lng2,lat2)`}
}
\description{
DISCLAIMER: the `_flex()` functions offer more flexibility in label placement and will offer new features under development for the `mapBliss` package. These functions will be subject to change more regularly than the standard functions.
}
\details{
DOCUMENTATION IS NEEDED

Plot multiple stops on a route `plot_route_flex()`. This function produces a html object so in order to save the image as an .svg (for printable visuals) check out the `save_map_svg()`.

For more information on how to integrate the MapBox API to make beautiful maps, check out [this blog](https://bensstats.wordpress.com/2021/10/25/robservations-16-using-the-mapbox-api-with-leaflet/).
}
\examples{

viz_1<-plot_hybrid_flex(addresses=c("Amsterdam","London","Newcastle upon Tyne"),
                        how=c("car","car"),
                        weight=3,
                        colour="red",
                        label_position = c("left","left","left"),
                        text_indent= c("-6em","-5em","-13em"),
                        mapBoxTemplate="//{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
                        )
viz_1
}
