% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_framed_map_svg.R
\name{save_map_frame_svg}
\alias{save_map_frame_svg}
\title{Save Framed Map as .svg}
\usage{
save_map_frame_svg(
  map,
  svg_name = "Rplot.svg",
  delay = 10,
  vwidth = 992,
  vheight = 1000
)
}
\arguments{
\item{map}{for `save_map_frame_svg()`. A framed map.}

\item{delay}{for `save_map_frame_svg()`. Delay (in seconds) before taking the mop screenshot. Set by default to 10 seconds.}

\item{vwidth}{for `save_map_frame_svg()`.}

\item{vheight}{for `save_map_frame_svg()`.}
}
\description{
Save framed map as .svg file for high quality printing!
}
\examples{
plot_city_view("Jersalem, IL") |>
 frame_1(title_text="Jerusalem", subtitle_text="City of Gold", subtitle_font="Brush Script MT")|>
 save_map_frame_svg()
## Deleting file for package speifics
file.remove("Rplot.svg")
}
