% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_map_svg.R
\name{save_map_svg}
\alias{save_map_svg}
\title{Save Map as .svg}
\usage{
save_map_svg(viz, svg_name = "Rplot.svg", zoom = 3)
}
\arguments{
\item{viz}{The viz you created with plot_route or plot_flights.}

\item{svg_name}{The name of the file you want to call}

\item{zoom}{How much you want to zoom in on your map. I find that setting `zoom=3` works best but try it out yourself to see what works.}

\item{map}{for `frame_*` functions. A framed map.}

\item{delay}{for `frame_*` functions. Delay (in seconds) before taking the mop screenshot. Set by default to 10 seconds.}

\item{vwidth}{for `frame_*` functions.}

\item{vheight}{for `frame_*` functions.}
}
\description{
Save map as .svg file for high quality printing!
}
\examples{


viz<- plot_route(c("Toronto","Niagara Falls","Monsey"),
                  how="car",
                  font="Courier",
                  label_position="right",
                  weight=1.5)

save_map_svg(viz)
## Deleting file for package speifics
file.remove("Rplot.svg")

plot_city_view("Jersalem, IL") |>
 frame_1(title_text="Jerusalem", subtitle_text="City of Gold", subtitle_font="Brush Script MT")|>
 save_map_frame_svg()
## Deleting file for package speifics
file.remove("Rplot.svg")
}
