% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frames.R
\name{frame_1}
\alias{frame_1}
\title{Put Map In A Custom Frame}
\usage{
frame_1(
  map,
  title_text = "Title",
  subtitle_text = "Subtitle",
  title_font = "Brush Script MT",
  subtitle_font = "Trebuchet MS",
  frame_width = "100\%",
  full_height = 1000,
  frame_height = 780
)
}
\arguments{
\item{map}{a map you created with mapBliss or leaflet}

\item{title_text}{Title text}

\item{subtitle_text}{Subtitle text}

\item{title_font}{Title font}

\item{subtitle_font}{Subtitle font}

\item{frame_width}{Set by default to 100 percent. It is recommended not to change this.}

\item{full_height}{The shiny screen. Set by default to 900}

\item{frame_height}{Set by default to 790 so as to fill the default shiny screen. Update according to your needs.}
}
\description{
Put your maps in a customized frame with `frame_*()`.
}
\details{
DOCUMENTATION IS NEEDED

`frame_1()` is a basic white square frame.
}
\examples{

plot_city_view("Jersalem, IL") |>
 frame_1(title_text="Jerusalem", subtitle_text="City of Gold", subtitle_font="Brush Script MT")
}
