% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_city_view.R
\name{plot_city_view}
\alias{plot_city_view}
\title{Plot A City View Map}
\usage{
plot_city_view(
  city,
  mapBoxTemplate = "//{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
  zoomControl = c(0.1, -0.1, 0.1, -0.1)
)
}
\arguments{
\item{city}{The city you want to get a view of.}

\item{mapBoxTemplate}{The MapBox template you want to use.}

\item{zoomControl}{The zoom control for your bounding box. Format is `c(lng1,lat1,lng2,lat2)`}
}
\description{
Plot a city view with `plot_city_view`. This essentially applies a bounding box to the MapBox template of your choosing to make a beautiful map of your favorite city!
}
\examples{

plot_city_view("New York City")

plot_city_view("Paris")
}
