% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flights.R
\name{plot_flights}
\alias{plot_flights}
\title{Plot Flight Paths}
\usage{
plot_flights(
  addresses,
  colour = "black",
  opacity = 1,
  weight = 1,
  radius = 2,
  label_text = addresses,
  label_position = "bottom",
  font = "Lucida Console",
  font_weight = "bold",
  font_size = "14px",
  text_indent = "15px",
  mapBoxTemplate = "//{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
  nCurves = 100
)
}
\arguments{
\item{addresses}{The locations which you want to plot your flight route. ORDER MATTERS so be sure to list the locations in the order you want to plot the flight path.}

\item{colour}{The colour of the points and route. By default set to "black"}

\item{opacity}{Opacity of the points and route on the map}

\item{weight}{Line thickness}

\item{radius}{Point size}

\item{label_text}{Alternative text to display as the address labels.}

\item{label_position}{where to place the label relative to the point}

\item{font}{font-family css property}

\item{font_weight}{font-weight css property}

\item{font_size}{font-size css property}

\item{text_indent}{text indent css property}

\item{mapBoxTemplate}{the MapBox template you want to use.}

\item{nCurves}{flight path smoothness. I have found setting this to 100 works best, but feel free to play around with it.}
}
\description{
Plot multiple flights with `plot_flights`. This function produces a html object so in order to save the image as an .svg (for printable visuals) check out the `save_map_svg()`.
}
\details{
For more information on how to integrate the MapBox API to make beautiful maps, check out [this blog](https://bensstats.wordpress.com/2021/10/25/robservations-16-using-the-mapbox-api-with-leaflet/).
}
\examples{
plot_flights(c("YYZ",
"GIG",
"CPT/FACT",
"BOM",
"TLV"),
weight = 1,
opacity = 1,
nCurves = 10,
label_position="right")

}
