% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hybrid_route.R, R/plot_hybrid_route_flex.R
\name{plot_hybrid_route}
\alias{plot_hybrid_route}
\title{Plot Hybrid Route}
\usage{
plot_hybrid_route(
  addresses,
  how = c("car", "flight", "bike", "foot"),
  colour = "black",
  opacity = 1,
  weight = 1,
  radius = 2,
  label_text = addresses,
  label_position = "bottom",
  font = "Lucida Console",
  font_weight = "bold",
  font_size = "14px",
  text_indent = "15px",
  mapBoxTemplate = "//{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
  nCurves = 100,
  zoomControl = c(0.1, 0.1, -0.1, -0.1)
)

plot_hybrid_route(
  addresses,
  how = c("car", "flight", "bike", "foot"),
  colour = "black",
  opacity = 1,
  weight = 1,
  radius = 2,
  label_text = addresses,
  label_position = "bottom",
  font = "Lucida Console",
  font_weight = "bold",
  font_size = "14px",
  text_indent = "15px",
  mapBoxTemplate = "//{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
  nCurves = 100,
  zoomControl = c(0.1, 0.1, -0.1, -0.1)
)
}
\arguments{
\item{addresses}{The stops on your route}

\item{how}{Routing profile(s) to use, e.g. "car", "bike" or "foot" (when using the routing.openstreetmap.de test server) AND "flight". n-1 routing profiles are required for for n addresses}

\item{colour}{what colour you want the route line to be colored}

\item{opacity}{line opacity - a value between 0 and 1}

\item{weight}{Line thickness}

\item{radius}{Point size}

\item{label_text}{Optional. Alternative text to display as the address labels.}

\item{label_position}{where to place the label relative to the point ("bottom", "top", "left", "right")}

\item{font}{font-family css property}

\item{font_weight}{font-weight css property}

\item{font_size}{font-size css property}

\item{text_indent}{text indent css property}

\item{mapBoxTemplate}{The MapBox template you want to use.}

\item{nCurves}{flight path smoothness. I have found setting this to 100 works best, but feel free to play around with it.}

\item{zoomControl}{The zoom control for your bounding box. Default Format is `c(lng1,lat1,lng2,lat2)`}
}
\description{
Plot multiple stops on a route `plot_hybrid route()`. This function produces a html object so in order to save the image as an .svg (for printable visuals) check out the `save_map_svg()`. This function differs from `plot_route()`  and `plot_flights()`  as it accomidates a hybrid route where some locations are traveled via flight and others are traveled via car, bike or foot.

DISCLAIMER: the `_flex()` functions offer more flexibility in label placement and will offer new features under development for the `mapBliss` package. These functions will be subject to change more regularly than the standard functions.
}
\details{
DOCUMENTATION IS NEEDED

Plot multiple stops on a route `plot_hybrid route_flex()`. This function produces a html object so in order to save the image as an .svg (for printable visuals) check out the `save_map_svg()`. This function differs from `plot_route_flex()`  and `plot_flights_flex()`  as it accommodates a hybrid route where some locations are traveled via flight and others are traveled via car, bike or foot.
}
\examples{
plot_hybrid_route(c("Detroit","Toronto","MKE"),
                 c("car","flight"))
plot_hybrid_route(c("Detroit","Toronto","MKE"),
                 c("car","flight"))
}
