% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hybrid_route.R
\name{plot_hybrid_route}
\alias{plot_hybrid_route}
\title{Plot Hybrid Route}
\usage{
plot_hybrid_route(
  addresses,
  how = c("car", "flight", "bike", "foot"),
  colour = "black",
  opacity = 1,
  weight = 1,
  radius = 2,
  label_text = addresses,
  label_position = "bottom",
  font = "Lucida Console",
  font_weight = "bold",
  font_size = "14px",
  text_indent = "15px",
  mapBoxTemplate = "//{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
  nCurves = 100
)
}
\description{
Plot multiple stops on a route `plot_hybrid route()`. This function produces a html object so in order to save the image as an .svg (for printable visuals) check out the `save_map_svg()`. This function differs from `plot_route()`  and `plot_flights()`  as it accomidates a hybrid route where some locations are traveled via flight and others are traveled via car, bike or foot.
}
\examples{
plot_hybrid_route(c("Detroit","Toronto","MKE"),
                 c("car","flight"))
}
