% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_route.R
\name{plot_route}
\alias{plot_route}
\title{Plot route}
\usage{
plot_route(
  to,
  from,
  how = "car",
  colour = "black",
  opacity = 1,
  weight = 1,
  radius = 2,
  label_text = c(to, from),
  label_position = "bottom",
  provider = providers$CartoDB.PositronNoLabels,
  font = "Lucida Console",
  font_weight = "bold",
  font_size = "14px",
  text_indent = "15px",
  saturation = 0,
  mapBoxTemplate = "//{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
)
}
\description{
Documentation is needed
}
\examples{

viz<- plot_route_Mapbox("Toronto","Niagara Falls",how="car",font="Courier",label_position="right",weight=1.5)

viz
}
