% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_map_svg.R
\name{save_map_svg}
\alias{save_map_svg}
\title{Save Map as .svg}
\usage{
save_map_svg(viz, svg_name = "Rplot.svg", zoom = 3)
}
\description{
Documentation is needed.
}
\details{
Save map
}
\examples{

#' viz<- plot_route("Toronto","
                  "Niagara Falls",
                  how="car",
                  font="Courier",
                  label_position="right",
                  weight=1.5)


save_map_svg(viz)
}
